get and put requests to chord_server don't block the gen_server.
the chord_server gen_server uses a thread to handle requests.
google the use of gen_server:reply/2 and you should find the pattern
that's used to turn gen_servers into non-blocking, threaded servers.
this led to a quirky (slightly non-obvious) implementation where if the
state of the gen_server needs to be modified by a spawned thread, it has
to call the gen_server with a 'change_state' message. Hisenbug issues
might be lurking there.